fix: change holders type from double to int in TokenDetailInfo model
- Updated the holders field in TokenDetailInfo to use int instead of double for better data accuracy.
- Adjusted related serialization and deserialization logic in generated files.
- Modified TokenInfoDisplay widget to reflect the change in holders type.

+++ lib/data/models/token_detail/info/token_state.dart
@@ -25,7 +25,7 @@ class TokenDetailInfo with _$TokenDetailInfo {
     @JsonKey(name: "market_cap") required double marketCap,
     @JsonKey(name: "liquidity") required double liquidity,
     @JsonKey(name: "volume_24h") required double volume24h,
-    @JsonKey(name: "holders") required double holders,
+    @JsonKey(name: "holders") required int holders,
     @JsonKey(name: "highest_increase_rate")
     @FlexibleStringConverter()
     String? highestIncreaseRate,
